literaturesuggest: make formdata_to_model pure
* Now it's side-effec free
* There's no need to pass workflow objects around
