+++ inspirehep/modules/literaturesuggest/templates/literaturesuggest/tickets/curation_core.html
@@ -30,8 +30,3 @@ Record url: {{ record_url|safe }}
 User comments:
   {{ user_comment|safe }}
 {% endif %}
-
-{% if references %}
-References:
-  {{ references|safe }}
-{% endif %}
