+++ inspirehep/modules/literaturesuggest/templates/literaturesuggest/tickets/curator_submitted.html
@@ -23,7 +23,3 @@
   User comments:
   {{ user_comment|safe }}
 {% endif %}
-{%- if references %}
-  References:
-  {{ references|safe }}
-{% endif %}
